+++ spring-intro/src/main/java/com/xiaoming/spring/aop_basic/traditional_impl/Calculator.java
@@ -1,11 +1,11 @@
 package com.xiaoming.spring.aop_basic.traditional_impl;
 
 public interface Calculator {
-    double add(double a, double b);
+    int add(int a, int b);
 
-    double sub(double a, double b);
+    int sub(int a, int b);
 
-    double mul(double a, double b);
+    int mul(int a, int b);
 
-    double div(double a, double b);
+    int div(int a, int b);
 }
